Minor change in scheduler modal to support improvements
